Fix for an issue with data table filtering and allowing for user defined filters (#175)
* Update DataTablesFiltering.cs

Fixed an issue when filtering on a data table throws an exception in cases there are no relevant columns to be filtered. Example: data table contains only numerical columns and a user types 'a letter'. Since the letter cannot be converted to a number, no where clause is generated and data.Where command was failing.

* Allowing for custom filter definition

Allowing for custom filter definitions by a user and fixing a bug in RegisterFilter method.

* Fix for EnumFilter

EnumFilter method was crashing when a value to search was not the exact enum value. This prevented any subsequent filters to be processed.